Allow booting builder after creation
