+++ docs/reference/buildx_create.md
@@ -12,6 +12,7 @@ Create a new builder instance
 | Name | Description |
 | --- | --- |
 | [`--append`](#append) | Append a node to builder instead of changing it |
+| `--bootstrap` | Boot builder after creation |
 | `--builder string` | Override the configured builder instance |
 | [`--buildkitd-flags string`](#buildkitd-flags) | Flags for buildkitd daemon |
 | [`--config string`](#config) | BuildKit config file |
